refactor(transform): extract super-detection helpers to satisfy file-size gate
call_inliner.rs grew past the 2000-line CI limit. Move the inline-recursion
guard and lexical-`super` detection helpers into a new super_detect.rs sibling
module, re-exported through mod.rs.
